Create styles inside every folder and move them for a better readability
