+++ src/components/Avatar.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 
-import { AvatarImg } from "../styles";
+import { AvatarImg } from "./styles";
 
 interface ContactAvatar {
 	avatar: string | undefined | React.ReactElement[];
