+++ src/components/Card.tsx
@@ -1,6 +1,6 @@
 import { Close } from "../assets/icons";
 import { useContacts } from "../context/ContactContext";
-import { CardContent, CardHeader } from "../styles";
+import { CardContent, CardHeader } from "./styles";
 
 export const Card = ({ children, title }: { children: React.ReactNode; title: string }) => {
 	const { deselectContact } = useContacts();
