+++ src/components/styles/avatar.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const AvatarImg = styled.img`
+	border-radius: 50%;
+	height: auto;
+	width: 32px;
+`;
