+++ src/components/styles/card.tsx
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const CardContent = styled.div`
+	background-color: white;
+	border-radius: 10px;
+	height: fit-content;
+	margin-left: 10px;
+	padding: 15px;
+`;
+
+export const CardHeader = styled.div`
+	align-items: center;
+	display: flex;
+	justify-content: space-between;
+
+	h5 {
+		color: #4e5661;
+		font-size: 14px;
+		font-weight: 600;
+	}
+`;
