+++ src/components/styles/contact.tsx
@@ -1,0 +1,25 @@
+import styled from "styled-components";
+
+export const NameWrapper = styled.div`
+	align-items: center;
+	display: flex;
+	justify-content: center;
+
+	p:first-child {
+		align-items: center;
+		margin-right: 6px;
+	}
+`;
+
+export const ContactWrapper = styled.div`
+	padding: 30px;
+	text-align: center;
+
+	img {
+		width: 100px;
+	}
+
+	.email {
+		margin-top: 0px;
+	}
+`;
