+++ src/components/styles/icons.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Icon = styled.svg`
+	color: #626262;
+	cursor: pointer;
+	height: 22px;
+	padding: 6px;
+	width: 22px;
+`;
