+++ src/components/styles/index.ts
@@ -1,0 +1,5 @@
+export * from "./avatar";
+export * from "./card";
+export * from "./contact";
+export * from "./icons";
+export * from "./table";
