+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 
 import { App } from "./App";
-import { AppRoot } from "./styles/StyledComponents";
+import { AppRoot } from "./styles";
 
 const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
 root.render(
