+++ src/sections/styles/contactsSection.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const ContactSectionWrapper = styled.section`
+	display: flex;
+
+	@media (max-width: 768px) {
+		display: flex;
+		flex-direction: column;
+		margin-top: 10px;
+	}
+`;
