+++ src/styles/global.tsx
@@ -1,0 +1,12 @@
+import { createGlobalStyle } from "styled-components";
+
+export const AppRoot = createGlobalStyle`
+	* {
+	font-family: Verdana, sans-serif;
+	}
+		
+	body {
+		background-color: #f3f3f3;
+	}
+    
+`;
