+++ src/styles/index.ts
@@ -1,1 +1,1 @@
-export * from "./StyledComponents";
+export * from "./global";
